requirements.txt
requests==2.22.0
pathlib==1.0.1
imgurpython==1.1.7
python-dotenv==0.10.3